Revert "add logo, loading dots, and overlay information in index-landing page…"
