+++ src/pages/index.tsx
@@ -1,43 +1,57 @@
-import Image from "next/image";
+import HeaderNav from "@/components/navigation/HeaderNav"
+import Nav from "@/components/navigation/NavBar"
+import Link from "next/link"
+import SpotifyPlayer from 'react-spotify-web-playback';
 import { useState, useEffect } from "react";
+import { authorize } from "./api/authorize";
 import { useRouter } from "next/router";
-import styles from '@/styles/Landing.module.css'
+import useRefreshToken from "@/hooks/useRefreshToken";
 
-export default function Landing() {
+export default function Home() {
 
-  const [loading, setLoading] = useState(true);
-    const router = useRouter();
+  const [accessToken, setAccessToken] = useState("");
+  const [codeVerifier, setCodeVerifier] = useState('')
 
-    useEffect(() => {
-		const handleComplete = () => {
-			setTimeout(() => {
-				setLoading(false);
-				router.push('/logIn');
-        console.log("loading...")
-			},2000)
-		};
+  const router = useRouter();
+  const code = router.query.code;
+  useRefreshToken(code as string);
 
-    handleComplete()
-    },[loading]);
-    
-    return (
-        <div className={styles.container}>
-            { loading &&  
-              <div className={styles.info}>
-                <Image
-                    src="/Logo/logo.png"
-                    alt="logo"
-                    height={354}
-                    width={360}
-                />
-                <div>
-                  <div className={styles.loading_dots}></div>
-                  <div className={styles.loading_dots}></div>
-                  <div className={styles.loading_dots}></div>
-                  <div className={styles.loading_dots}></div>
-                </div>
-              </div>
-            }
+  useEffect(() =>{
+    setAccessToken(`${localStorage.getItem("access_token")}`)
+    console.log(accessToken)
+
+  })
+
+  return (
+    <main className={``} >
+      <HeaderNav text="Welcome back, John" type="profile"/>
+      <div id="mainContainer" className={`flex flex-col`}>
+        <h1>This is the index/home page</h1>
+        <Link href='/logIn'>Go to login</Link>
+        <Link href='/browse'>Go to browse </Link>
+        <Link href='/allNeighbourhood'>Go to allNeighbourhood</Link>
+        <Link href='/neighbourhood/[area]'>Go to neighbourhood area page</Link>
+        <Link href='/playMusic'>Go to playMusic</Link>
+        <Link href='/playArt'>Go to playArt</Link>
+
+        <p>Web player test</p>
+        <button onClick={authorize}>Authorize</button>
+        <div>
+          {
+            accessToken ?
+
+              <SpotifyPlayer
+                token={accessToken}
+                uris={['spotify:artist:6HQYnRM4OzToCYPpVBInuU']}
+              />
+              :
+              <p>No access token</p>
+          }
+
         </div>
-    )
+
+      </div>
+      <Nav type="home" />
+    </main>
+  )
 }
