+++ src/pages/logIn.tsx
@@ -4,8 +4,7 @@ export default function LogIn() {
   return (
     <main className={`flex min-h-screen flex-col items-center justify-between p-24`} >
       <div id="mainContainer" className={`flex flex-col`}>
-        <p>this is the login screen</p>
-        <Link href='/home'><b>Go to home</b></Link>
+
       </div>
     </main>
   )
