--- src/pages/overlayInformation.tsx
@@ -1,18 +1,0 @@
-
-
-export default function overlayInformation() {
-    return (
-    
-      <main className={``}>
-        <h1>Justin Langlois</h1>
-        <p>Title: Should I Be Worried?</p>
-        <p>Year: 2017</p>
-        <p>Ownership: City of Vancouver</p>
-        <p>Neighbourhood: Mount Pleasant</p>
-        <p>Status: In Place</p>
-        <p>This public artwork was produced by local artist Justin Langlois as part of the City’s first Artist-in-Residence program. Since mid-2016, Justin has been working with the Sustainability Group, learning about the various initiatives and themes that shape the Greenest City Action Plan and contributing to ideas and opportunities for greater engagement. The work produced aims to open up an ongoing dialogue about social, economic, and environmental sustainability.This project is supported by the City of Vancouver through the Public Art Program and the Sustainability Group with assistance from Engineering Services, Utilities, Streets and Electrical Design and Operations.</p>
-      </main>
-   
-    )
-  }
-
