--- src/styles/Landing.module.css
@@ -1,53 +1,0 @@
-.container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-}
-
-.info {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.loading_dots {
-    height: 15px;
-    width: 15px;
-    transform: scale(0.2);
-    border-radius: 50%;
-    display: inline-block;
-    margin: 0.5rem;
-    animation: scaling 1.5s ease-in-out infinite;
-}
-
-@keyframes scaling {
-    0%, 100% {
-        transform: scale(0.2);
-        background-color: var(--berry);
-    }
-    40% {
-        transform: scale(1);
-        background-color: var(--berryHover);
-    }
-    50% {
-        transform: scale(1);
-        background-color: var(--berryHover);
-    }
-}
-
-.loading_dots:nth-child(0) {
-    animation-delay: 0s;
-}
-.loading_dots:nth-child(1) {
-    animation-delay: 0.125s;
-}
-.loading_dots:nth-child(2) {
-    animation-delay: 0.25s;
-}
-.loading_dots:nth-child(3) {
-    animation-delay: 0.375s;
-}
-.loading_dots:nth-child(4) {
-    animation-delay: 0.5s;
-}
